feat(util-dev): add DockerUtil (hoisted from cardano-services tests)
