+++ packages/util-dev/package.json
@@ -65,7 +65,10 @@
   "dependencies": {
     "@cardano-sdk/core": "^0.9.0",
     "@cardano-sdk/util": "^0.8.1",
+    "@types/dockerode": "^3.3.8",
     "delay": "^5.0.0",
+    "dockerode": "^3.3.1",
+    "dockerode-utils": "^0.0.7",
     "envalid": "^7.3.1",
     "get-port-please": "^2.5.0",
     "json-bigint": "^1.0.0",
